requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
scrapy==2.11.1
google-generativeai==0.4.1
slowapi==0.1.9
structlog==24.1.0
python-dotenv==1.0.1
pydantic==2.6.4
requests==2.31.0
beautifulsoup4==4.12.3
